pooled variant detection improvements
* enable use-best-n-alleles

The --use-best-n-alleles parameter was previously disabled for non-SNP
variation, but this prevents its use as a bound on computational complexity.
This occurs readily in the case of multiple --ploidy 20 pools and
--pooled-discrete, leading to an combinatorial explosion in possible genotype
states.  A common result of --pooled-diploid would be the exhaustion of system
memory.

Users can now safely use --pooled-discrete provided they also use a suitable
setting for --use-best-n-alleles.  In practice, setting to 5 or lower should be
sufficient to prevent memory blowup in most situations.  For the time being, I
suggest testing with progressively lower settings, or simply setting it as low
as you think reasonable.  (For pooled experiments focused on SNPs, this would
be 2.)

This update includes two other fixes:

* uppercase reference sequence

Uppercasing the reference allele properly resolves an error with
FastaReference::getSubSequence (negative length).

* no hwe priors for pooled samples

The HWE component of the mappability estimate should be turned off in the case
of pooled-discrete runs, so it is now turned off when --pooled-discrete is
specified.