correctly retain observations during partial observation generation, allele balance fix
A coding error resulted in observation loss after assembling partial observations.

Limit the "probe" length required of all observations when approximating
correct allele balance for indels (specifically insertions).  This limits the
probe length to 50bp, not-configurable.  It will be updated to a measure based
on the average read length for future-correctness.