resolve "...is out of order! expected after..." bug
This error message was generated when attempting to gather partial support for
haplotypes.  If an alignment did not span the haplotype (which is unlikely,
because the haplotype calls tend to be short, but happens with reads that are
heavily soft-clipped), then this warning would be triggered and the read would
be ignored.  The adjustment allows for the use of such alignments when
generating partial support.