indel genotye likelihoods
This commit resolves a number of issues with indel genotype likelihoods.

First, although haplotype detection has proceeded across repeat structures in
the reference, the same extension was not applied to repeats which were not
represented in the reference, and only in reads.  This allowed reads supporting
the alternate to appear as if they supported the reference, leading to false
heterozygote calls for homozygous alternates.

Second, in some cases, it may be desireable to call haplotypes across regions
of low complexity.  --min-repeat-entropy provides this facility, requiring a
given number of bits per base of any reference-relative haplotype window built
around a repeat structure.

Thirdly, improvements in performance were provided by careful handling of
partial observations, specifically checking if apparent full-length haplotype
observations in fact support alternate haplotypes,

Fourthly, "null" observations (portions of reads which are N, variants that are
not specified in --haplotype-basis-alleles), are now correctly treated as
non-observations.  Previously, these were incorrectly coerced to be reference.

In all, these changes yield a 2% improvement in the area under the curve for
detection in 100 simulated 10x genomes (0.917 -> 0.937).  They also eliminate a
number of pathological errors in 1000 Genomes.

Also!  This commit includes bugfixes  for invalid memory access errors detected
by valgrind.