ignore un-flanked deletions at the beginning and end of alignments
Some aligners will report deletions at the beginning of alignments.  It is not
clear how to interpret these cases because, without flanking sequence in the
read, the indication of a relative deletion in the read is meaningless.  For
instance, a cigar for a 70bp read could be 1D70M.  What base in the read has
been deleted?

This resolves some bugs which generate errors like this:

    deletion...
    alt is empty