version 0.9.10
This begins the v1 release candidate series.  The change in version numbering
style recognizes the need to move the version past 0.9.9!  However, version 1.0
is meant to coincide with paper publication or submission.  Soon!

A large number of small changes land in this version, including changes to the
genotype likelihood calculations:

* Support for correct genotyping of large deletions.  This fix involved
correcting handling of "partial" observations of the reference allele.

* Default use of mapping quality in genotype likelihood calculations (this is
disabled via the --standard-gls flag).  Mapping qualities are incorporated via
the Effective Base Depth metric from snpTools (Baylor).

* Exclusion of mapping quality 0 mappings.  (This can be disabled with
--min-mapping-quality -1 --standard-gls).  The change in genotype likelihood
calculations means these are not considered meaningful.

These three changes improved the ROC-AUC for indels in simulation by about 1%,
which is substantial given that it now stands at 0.95.  No difference was
recorded for SNPs.

The commit also includes a change relevant to the output of variant records.

* Removal of variant parsing routines meant to standardize VCF records by
removing redundant information from the ALT/REF pairs.

These manipulation routines would, for example, remove portions of the REF and
ALTs which were always matching.  The canonical case would be a haplotype call
in which only a SNP was ultimately called.  The call would occur over several
bp of reference.  The manipulation is complex and caused errors in many
instances, which leads to many maintenance issues and sad users.

At present, VCF normalization is best-handled by external utilities, such as
vcfallelicprimitives from vcflib.  Retaining the haplotype information in the
call in this way clarifies which calls were attempted haplotype calls and which
weren't, regardless of output state.