resolve bug in implementation of Ewens' Sampling Formula
Thanks to Severine Catreux for the catch.

This should have the largest effect on sites that are multiallelic.  The
quality of these will be diminished slightly, in line with the bug.  This may
reduce the overall power to detect variants at multiallelic loci.

It is possible that this is not the most-correct way to utilize the ESF in
freebayes, as the assumption of a constant mutation rate for each locus in the
genome is inadequate.  At some loci the mutation rate is orders of magnitude
higher than elsewhere.  Problematically, these are the same places where
Illumina data (and any PCR-based library) tends to present artifacts.