swap default GL calculations; add --experimental-gls (replaces --standard-gls)
A number of reports have recently described issues with freebayes genotyping.
These generally appear to stem from the new genotype likelihood calculations
which were implemented and tested primarily in the context of re-genotyping
putative alleles in diploids for the purpose of generating high-quality GLs
suitable for imputation.

Testing has shown that in most cases this change has little effect, but where
it does matter is also where it could be very confusing for users, such as in
polyploid organisms or in low-complexity sequence (where the "experimental" GL
calculation likes to call spurious hets).

The default (previously --standard-gls) is restored.  Now, to obtain the new
behavior you should supply --experimental-gls or provide contamination
estimates or a hard setting of contamination via --prob-contamination.