improve handling of --variant-input
This feature has been mostly broken for some time, superseded instead by
--haplotype-basis-alleles, which targets haplotype construction and genotyping
only to a particular set of alleles.  However, the logic of the haplotype
construction requires observations to support putative haplotype structures,
emitting no information about sites that are called homozygous reference.

To resolve this issue, I have fixed blocking issues for --variant-input, and
now suggest its use in the cases that:

- you want to be sure to make calls at given sites
- you are calling a number of samples or pools singly, and want to merge them
on a given set of known alleles
- you otherwise want to directly interrogate a given BAM file at given positions

--variant-input and --haplotype-basis-alleles are not quite harmonized yet.  I
am still thinking about the correct way to do this.  Nevertheless, it should
now be possible to use the feature as was intended.

To do so, normalize your variant representation by e.g. vt normalize (from
https://github.com/atks/vt).  Remove sample information and allele frequency
(AF) and count (AC) from the input file.  These (AC, AF, GT, GL) can be left if
desired, but they are used as priors when computing the site.  Then tabix index
the VCF and provide it to the algorithm via the --variant-input flag.

If you would like to restrict to only variant-input sites, also supply
--only-use-input-alleles.  Otherwise, calls will be made as normally at sites
and alleles that aren't in the variant-input file.