correct buildHaplotypeAlleles, and clarify --min-mapping-quality
Due to a logic error, buildHaplotypeAlleles was excluding reads that end
exactly at the haplotype window.  Many thanks to Severine Catreux, who made
this catch.

In this commit, -m is changed to (as documented) exclude reads with MQ < the
given value.  It previously was <=.  The default -m is set to 1.  This will
affect users by adjusting the minimum cutoff, which they should increase by 1
for equivalent behavior.