freebayes-parallel
This script depends on GNU parallel, but allows for freebayes to be run with
near-100% efficiency in an arbitrary number of threads.  Two positional
arguments are required, $regions and $ncpus; the rest of the command line can
be exactly identical to what is typically used in freebayes with the exception
of the output parameter -v (VCF), which cannot be used.

Parallelization requires a list of regions, which can be generated inline in
one command as shown in the help text of the executable (using
fasta_generate_regions.py), or can be generated from an arbitrary BAM file
using coverage_to_regions.py, which is also in the scripts directory.  The
second option may provide lower variance in sub-job runtime and thus better
throughput at the cost of having to traverse an entire alignment file to
generate the regions.