build tracking
Per requirements of our funding source (NIH), we need to track usage of
freebayes.  As a first step, each 'make' invocation will grab a URL which
encodes the freebayes version.  The logs will be parsed to understand the
number of unique builds and users.

If this bothers you, simply remove the second line (wget -q ...) in the root
Makefile.

Before doing so, please consider how helpful this information is to our
continued support of the method.  We will use this information only to
determine the number of unique users and rough geographical distribution, as
these are important for funding.