resolve an overly-cautious bounds check and allow really long haplotypes!
This removes an apparently extraneous check of basesRight in haplotype
construction which appears to limit the length of haplotypes that can be
described to a fraction of read length.

This may have caused some quite serious problems for those attempting long
haplotyping.

This commit is provisional, it may be reverted before being pushed to the
public repo if it is found to be problematic via regression tests.