reaolve --varoant-input issues
Fixes bug with deletion handling. Thanks to Sebastian Waszak for identifying
the issue.

Disables AF, genotype, and GL-based priors. These will be reenabled and
directly triggered by new flags in a following commit.

This hidden functionality is not working correctly and has caused problems for
many users because the freebayes output would not function as input to the
method. As of this commit, it will without additional manipulation, field
stripping, or normalization.