represent GLs in full dynamic range
This can be really important when we are working with somatic calls.

There is no reason that I know of to cap the GL representation, although it
should help some with compression in data that is predominantly GLs.

I haven't instrumented an option to enable this from the command line, but the
capping is controlled by Parameters.limitGL. When set to 0, it uses the full
dynamic range. When set to any other negative value (in log space) it bounds so
that value.