resolve #109 don't always pick the "best het" combo
As freebayes used to be used solely in a discovery context, the maximum
posterior genotyping including any non-reference alleles has been the default
output of freebayes.

The quality scores of such sites tend to be very low, and so they are often
filtered out and ignored. This has protracted the confusion, however, because
in targeted genotyping you want to know the precise genotype for every sample.

It is better to output the MAP genotyping regardless of whether or not it has
variation in it. This simplifies a number of use-cases and is more
straightforward, matching users' expectations about how the method should
perform.