resolve #98: allow long haplotype detection with --haplotype-basis-alleles
In the past, --haplotype-basis-alleles was used for targeted genotyping in e.g.
1000G phase 3. A number of problems were shown to arise when allele
observations not in the haplotype-basis input were were forced into reference
space, so I made a change a year ago that treated these as if they were
uncalled bases, or Ns (ALLELE_NULL in freebayes internal terminology).

However, this introduced a new issue, which is that long haplotype detection
over a targeted set of alleles was now broken, as the null allele breaks the
haplotype in the semantics of the system.

At the same time, improvements in --variant-input now mean that this method can
be used for targeted genotyping. The haplotype-basis method is now free to be
used specifically for the contiguous phasing of local regions based on a set of
input alleles. This change resets the algorithm to behave as it did previously,
treating allele observations not in the --haplotype-basis-alleles input set as
if they are actually observations of the reference. For most applications of
this type, this is not an issue. Those wishing to run targeted genotyping at
particular alleles are encouraged to use --variant-input instead, which does
not attempt to build haplotype observations across the reads.