Add debugging for mergeAllele() callsites in addAllele()
The messages helped me to understand and verify just how the allele
management happens.

I put the debugging at the callsites instead of in mergeAllele() itself
so that the messages can, for example, identify *which* mergeAllele()
call in a local scope and include more than one previous allele when
useful.

The first callsite (mergeAllele/1) is DEBUG2 because it's likely to be
orders of magnitude more frequent and less useful than the other
callsites.

It's a little gross to pass in the Parameters object to
RegisteredAlignment's constructor, but a better way wasn't obvious to
me.