Respect maxComplexGap when reference alleles are added in small lengths
If a non-reference allele would extend a complex allele, the last part
of the complex allele must be either a non-match or a match of less than
the maxComplexGap (-E / --haplotype-length).  This prevents the alleles
(1X, 1M, 1M, 1M, 1X), when sequentially added with addAllele(), from
being merged into a complex allele 1X3M1X if the maxComplexGap is less
than 3.  The case of sequentially adding the alleles (1X, 3M, 1X) is
already handled in an earlier conditional block which checks if
newAllele.referenceLength > maxComplexGap.

In the case when maxComplexGap would be violated, the complex allele is
split in two, with the second part being the final embedded match, and
the new allele (a non-reference allele) is simply pushed onto the list.