output max(GL) genotypes over the alleles in the maximum posterior gentoyping
This is a bit complex, so hold on.

I noticed that in practice, the best-quality results came from running
freebayes and then setting the called genotypes to the maximum GLs post-call
using vcfglxgt. This seemed needlessly complex for a standard process, so I
implemented the same functionality in the main loop.

With this change, the default output works as follows:

1. a maximum posterior genotyping is established through the standard EM-like
posterior search procedure
2. the potential genotypes are subset those composed of alleles in the MAP
genotyping
3. the reported genotypes are assigned as the maximum-likelihood gentoype among
those that are composed of the alleles in the MAP genotyping

The result is as described in the title, we always report the max(GL) over the
alleles in the maximum posterior genotyping.

If you're using --report-genotype-likelihood-max, you may be very interested in
this commit. I believe it may not be necessary to do so anymore, and in fact it
may be undesireable, as the results from this setting are subtly different.
They return the GL max over *all putative alleles*, which is probably not what
you really wanted anyway.