Revert "remove +1 bounds check on coordinates"
This reverts commit 22295cfc7f3edb1f687e261457cc5c5705062919.

The right fix for this is actually a few lines above:

    (stopPos == -1) ? reference.sequenceLength(startSeq) : stopPos - 1); // internally, we use 0-base inclusive end

should be

    (stopPos == -1) ? reference.sequenceLength(startSeq) - 1 : stopPos - 1); // internally, we use 0-base inclusive end
or
    ((stopPos == -1) ? reference.sequenceLength(startSeq) : stopPos) - 1); // internally, we use 0-base inclusive end

and then this commit should be reverted. Other targets like chrZ:0-200
are already adjusted to be fully-closed in targets, so the correct
comparison is bd.right + 1 > reference.sequenceLength(...). The bug is
that targets like chrZ were left half-open by mistake.

I missed this when I was auditing!