use "experimental-gls" by default
For a long time, the best way to genotype diploids using freebayes was by
specifying --experimental-gls.  This model incorporates a lot of things into
the genotype likelihoods (GLs) that matter, such as the effective base depth
metric (Q = MQ * BQ) from SNPTools, and partial observation support (which
matters for big deletions, and long haplotypes).

However, bugs when handling polyploids prevented me from setting it as default.

With this commit, these issues are resolved and the default is now set to what
was previously specified by "--experimental-gls".

Use --legacy-gls for old polybayes-equivalent operation.