Prevent word splitting in freebayes-parallel args
You might want to call freebayes-parallel with filenames with spaces
in them:

freebayes-parallel 'a regions file.txt' 2 -f 'some reads.fasta' alns.bam

But, when the argument array is assigned to the command string, everything
gets split on spaces, and freebayes would do something like complain about
not being able to find a file called "some". This change keeps the arguments
in an array and prevents splitting of the regionsfile.

Two tests are included with filenames that have spaces in them.