Add tests for variant input
The current test for the -@ parameter is this:

is $(freebayes -f tiny/q.fa -@ tiny/q.vcf.gz tiny/NA12878.chr22.tiny.bam | grep -v "^#" | wc -l) 19

but this also passes without the variant input flag:

is $(freebayes -f tiny/q.fa tiny/NA12878.chr22.tiny.bam | grep -v "^#" | wc -l) 19

So, it's not really constraining the behavior of -@ very much. This commit
adds two test that use a vcf with fake variants at q:500, q:1000, and q:11000
and checks that they appear in the output vcf or that only they appear in the
output vcf when -l is used.