Add variant input tests with -r and -c
Presently, if you set a region with -r and a variants input file with -@, the
variant inputs will be ignored. These tests confirm that, and when they pass
will confirm a fix. In addition, the method responsible for this behavior also
does special things when reading the BAM from stdin, so this change adds some
test coverage for that as well.