Fix variant inputs when region is specified
When a variant input file is given with -@,
AlleleParser::usingVariantInputAlleles is set to true. This makes AlleleParser
look for variants in AlleleParser::inputVariantAlleles. But this is only
populated if AlleleParser::targets is empty, which is only true if a region
is not specified. This change populates the inputVariantAlleles map even when
targets is not empty.