add gVCF support
When --gvcf is supplied, freebayes will output blocks for each
region that does not have a call. These have a particular format
that is not exactly the same as other records. A genotype quality
is provided, but this is just the reference quality sum adjusted
by the alternate quality sum. These are provided per-sample.
Average depth is provided across the region. Improvement of this
format is essential, and this version should stand as a
placeholder for future development!