replace FORMAT/DPR annotation by FORMAT/AD
DPR is mislabeled as a `Number=A` field, but writing out `Number=R` values.
Also, as @nh13 noticed, the REF allele count was wrong anyway -
printing the total depth rather than the REF depth.

Replace by FORMAT/AD which is a Number=R field now defined in the spec. Also
need to bump the VCF version to 4.2 (`Number=R` was introduced in 4.2, but AD in 4.3)

Fixes #267